+++ src/main/java/lazyalienserver/carpetlasaddition/CarpetLASClient.java
@@ -1,19 +1,38 @@
 package lazyalienserver.carpetlasaddition;
 
+import com.mojang.brigadier.CommandDispatcher;
+import lazyalienserver.carpetlasaddition.commands.Client.BinaryCommand;
+import lazyalienserver.carpetlasaddition.commands.Client.CalculateCommand;
+import lazyalienserver.carpetlasaddition.commands.Client.CalcPearlCommand;
 import lazyalienserver.carpetlasaddition.commands.Client.SetRenderDistanceCommand;
+import lazyalienserver.carpetlasaddition.commands.Client.DecimalCommand;
 import lazyalienserver.carpetlasaddition.commands.TestCommand;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.command.v1.ClientCommandManager;
+import net.fabricmc.fabric.api.client.command.v1.FabricClientCommandSource;
 
 @Environment(EnvType.CLIENT)
 public class CarpetLASClient implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
-        TestCommand.register(ClientCommandManager.DISPATCHER);
-        SetRenderDistanceCommand.register(ClientCommandManager.DISPATCHER);
-        //CalcPearlCommand.register(ClientCommandManager.DISPATCHER);
+        registerCommand(ClientCommandManager.DISPATCHER);
     }
+
+    public static void registerCommand(CommandDispatcher<FabricClientCommandSource> dispatcher){
+        //Calc
+        BinaryCommand.register(dispatcher);
+        DecimalCommand.register(dispatcher);
+        CalculateCommand.register(dispatcher);
+
+        //Setting
+        SetRenderDistanceCommand.register(dispatcher);
+        CalcPearlCommand.register(dispatcher);
+
+        //Test
+        TestCommand.register(dispatcher);
+    }
+
 }
