+++ src/main/resources/assets/carpet-las-addition/lang/en_us.json
@@ -12,14 +12,5 @@
   "rule.CompostGradeModification.desc": "Any item placed in the bin will add one level to the bin [you need to right-click the bin once after each setting change to apply the current setting]",
 
   "rule.MinecartDropModify.name": "minecart drop mechanism modified",
-  "rule.MinecartDropModify.desc": "Replace the minecart drop mechanism with version 1.18.2 with the drop mechanism in version 1.19",
-
-  "rule.commandCalculate.name": "commandCalculate",
-  "rule.commandCalculate.desc": "A Calculate Command in minecraft",
-
-  "rule.commandBinary.name": "commandBinary",
-  "rule.commandBinary.desc": "A Command,it can Decimal to Binary",
-
-  "rule.commandDecimal.name": "commandDecimal",
-  "rule.commandDecimal.desc": "A Command,it can Binary to Decimal"
+  "rule.MinecartDropModify.desc": "Replace the minecart drop mechanism with version 1.18.2 with the drop mechanism in version 1.19"
 }
