+++ src/main/resources/assets/carpet-las-addition/lang/zh_cn.json
@@ -12,14 +12,5 @@
   "rule.CompostGradeModification.desc": "任何放入堆肥桶的物品都将使堆肥桶增加一级[每次更改设置后需要右键一次堆肥桶才能应用当前设置]",
 
   "rule.MinecartDropModify.name": "矿车掉落机制修改",
-  "rule.MinecartDropModify.desc": "将1.18.2版本的矿车掉落机制替换成1.19版本的矿车掉落机制",
-
-  "rule.commandCalculate.name": "计算指令",
-  "rule.commandCalculate.desc": "添加了计算指令",
-
-  "rule.commandBinary.name": "转二进制指令",
-  "rule.commandBinary.desc": "将十进制转换为二进制",
-
-  "rule.commandDecimal.name": "转十进制指令",
-  "rule.commandDecimal.desc": "将二进制转换为十进制"
+  "rule.MinecartDropModify.desc": "将1.18.2版本的矿车掉落机制替换成1.19版本的矿车掉落机制"
 }
